refactor(api): split matrix.js god-module — extract builders + saved filters (Q1)
matrix.js was 1,679 lines. First, behavior-preserving slice of the audit's Q1:

- Move the 5 pure roll-up SQL builders (buildRollupSql, buildRollupRolesSql,
  buildRolesAsRowsSql, buildGroupTotalsSql, buildRolesDrillSql) to
  matrix/rollupBuilders.js; matrix.js imports them. Their unit tests move to
  matrix/rollupBuilders.test.js (co-located with the module).
- Move the saved-filter CRUD + org-wide default-filter endpoints to
  routes/matrix/savedFilters.js, mounted via router.use() so public paths are
  unchanged (/api/matrix/saved-filters, /api/matrix/default-filter).

matrix.js 1,679 -> ~1,415 lines. Pure code move: no query logic, route paths,
request/response shapes, or ordering changed. The 680-line /matrix/data handler
and the scope endpoints are intentionally left for a follow-up slice.

Validated on sidekicks (SK1 small data, SK3 large real data) before/after:
vitest green, stack boots clean, saved-filter CRUD works, and matrix
scope-stats + roll-up responses are byte-identical to the latest-main baseline.
