+++ changes/split-matrix-module.md
@@ -1,0 +1,1 @@
+- Internal maintainability: split the oversized matrix API route module into smaller focused modules — the saved-filter endpoints and the roll-up SQL builders now live in their own files. No change to matrix behaviour, endpoints, or output.
